+++ xpumd/docs/CHANGES.md
@@ -7,6 +7,8 @@
     * **Action required:** `hw_status{hw_type="gpu",hw_state="ecc_<state>"}` queries need to be updated to `hw_gpu_ecc_state{hw_state="<state>"}`
   * `hw.memory.ecc` attribute of the `hw.gpu.info` metric renamed to `hw.gpu.ecc.support` (to match the new ECC state metric)
     * **Action required:** `hw_gpu_info{hw_memory_ecc=...}` queries need to be updated to `hw_gpu_info{hw_gpu_ecc_support=...}`
+* XPUMD metrics: New `ok` state for the GPU device in the `hw.status` metric
+  * `hw_status{hw_type="gpu",hw_state="ok"}` aggregates the device level GPU states.
 
 
 ## XPUM 2.2.0